refactor: gather all file operations to a file
