+++ kudu/commands/init.py
@@ -3,7 +3,7 @@ import os
 import json
 from kudu.api import request
 from kudu.config import write_config
-
+from kudu.file import create_file
 
 @click.command()
 @click.pass_context
@@ -19,41 +19,6 @@ def init(ctx):
 
     write_config({'file_id': file_id})
 
-
-def create_file(token, app_id, file_body, download_url = None, file_data= None):
-    requestJSON = {
-        'app':
-            app_id,
-        'body':
-            file_body
-    }
-
-    if download_url:
-        requestJSON['downloadUrl'] = download_url
-
-    if file_data:
-        files = {
-            'json': (None, json.dumps(requestJSON), 'application/json'),
-            'file': (os.path.basename(file_body), file_data, 'application/octet-stream')
-        }
-
-    res = request(
-        'post',
-        '/files/',
-        files=files,
-        token=token
-    )
-
-    resJson = res.json()
-
-    if res.status_code != 201:
-        if resJson.get('app'):
-            click.echo('Invalid instance', err=True)
-        else:
-            click.echo('Unknown error', err=True)
-        exit(1)
-
-    return resJson.get('id')
 
 
 def validate_file(file_id, token):
